src/main/java/lexicon/fundamentals/oop/Customer.java
package lexicon.fundamentals.oop;

public class Customer {
    private int customerId;
    private String customerFirstName;
    private String customerLastName;
    private String customerEmail;

    public Customer(int customerId, String customerFirstName, String customerLastName, String customerEmail) {
        this.customerId = customerId;
        this.customerFirstName = customerFirstName;
        this.customerLastName = customerLastName;
        this.customerEmail = customerEmail;
    }

    public void setCustomerId(int customerId) {
        this.customerId = customerId;
    }

    public void setCustomerFirstName(String customerFirstName) {
        this.customerFirstName = customerFirstName;
    }

    public void setCustomerLastName(String customerLastName) {
        this.customerLastName = customerLastName;
    }

    public void setCustomerEmail(String customerEmail) {
        this.customerEmail = customerEmail;
    }

    public int getCustomerId() {
        return customerId;
    }

    public String getCustomerFirstName() {
        return customerFirstName;
    }

    public String getCustomerLastName() {
        return customerLastName;
    }

    public String getCustomerEmail() {
        return customerEmail;
    }

   public String getCustomerInformation(){
        return "Customer ID: "+getCustomerId()+" Customer Name: "+getCustomerFirstName()+" "+getCustomerLastName()+" Customer Email: "+getCustomerEmail();
   }


}
